Refactor add Package
