+++ .idea/misc.xml
@@ -8,6 +8,7 @@
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.25317028985507245" />
         <entry key="app/src/main/res/layout/activity_test.xml" value="0.18958333333333333" />
         <entry key="app/src/main/res/layout/fragment_expiration_date.xml" value="0.11043566362715299" />
+        <entry key="app/src/main/res/layout/fragment_note.xml" value="0.29074074074074074" />
         <entry key="app/src/main/res/layout/fragment_search.xml" value="0.20643363728470113" />
         <entry key="app/src/main/res/layout/item_recipe.xml" value="0.1828774062816616" />
       </map>
