+++ app/src/main/java/com/example/ingredient/activity/LoginActivity.kt
@@ -1,4 +1,4 @@
-package com.example.ingredient
+package com.example.ingredient.activity
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -13,6 +13,8 @@ import com.android.volley.Request
 import com.android.volley.Response
 import com.android.volley.toolbox.JsonObjectRequest
 import com.android.volley.toolbox.Volley
+import com.example.ingredient.R
+import com.example.ingredient.network.SessionCallback
 import com.facebook.AccessToken
 import com.facebook.CallbackManager
 import com.facebook.FacebookCallback
