+++ app/src/main/java/com/example/ingredient/activity/MainActivity.kt
@@ -1,25 +1,21 @@
-package com.example.ingredient
+package com.example.ingredient.activity
 
-import android.app.Activity
-import android.app.Application
 import android.graphics.Rect
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.view.MotionEvent
-import android.view.View
 import android.view.inputmethod.InputMethodManager
-import android.widget.RadioButton
 import androidx.fragment.app.Fragment
-import com.example.ingredient.database.ExpirationDateDatabase
-import com.example.ingredient.fragment.ExpirationDate
-import com.example.ingredient.fragment.FoodBook
-import com.example.ingredient.fragment.Search
-import com.example.ingredient.fragment.Note
+import com.example.ingredient.R
+import com.example.ingredient.data.ExpirationDateDatabase
+import com.example.ingredient.src.ExpirationDate
+import com.example.ingredient.src.FoodBook
+import com.example.ingredient.src.search.Search
+import com.example.ingredient.src.Note
 import com.example.ingredient.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
-    private lateinit var searchFragment:Search
+    private lateinit var searchFragment: Search
     private lateinit var foodbookFragment:FoodBook
     private lateinit var expirationdateFragment:ExpirationDate
     private lateinit var noteFragment:Note
@@ -42,7 +38,7 @@ class MainActivity : AppCompatActivity() {
         // 초기 화면을 Search 프래그먼트로 설정
         val transection = supportFragmentManager
         transection.beginTransaction().
-            add(R.id.fragment_container,Search(),"Search").commit()
+            add(R.id.fragment_container, Search(),"Search").commit()
 
         // 하단바를 통해 화면(프래그먼트) 전환
         binding.menuBottom.setOnItemSelectedListener { id ->
@@ -78,9 +74,9 @@ class MainActivity : AppCompatActivity() {
                     getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager
                 if (imm != null)
                     imm.hideSoftInputFromWindow(focusView.getWindowToken(), 0);
-                focusView.clearFocus();
+                focusView.clearFocus()
             }
         }
-        return super.dispatchTouchEvent(ev);
+        return super.dispatchTouchEvent(ev)
     }
 }
