+++ app/src/main/java/com/example/ingredient/src/ExpirationDate.kt
@@ -1,6 +1,5 @@
-package com.example.ingredient.fragment
+package com.example.ingredient.src
 
-import android.app.Application
 import android.os.Build
 import android.os.Bundle
 import android.util.Log
@@ -8,11 +7,8 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.RadioButton
 import androidx.annotation.RequiresApi
-import com.example.ingredient.R
-import com.example.ingredient.database.ExpirationDateDao
-import com.example.ingredient.database.ExpirationDateDatabase
+import com.example.ingredient.data.ExpirationDateDatabase
 import com.example.ingredient.databinding.FragmentExpirationDateBinding
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -35,7 +31,7 @@ class ExpirationDate : Fragment() {
             var text = binding.editTest.text.toString()
             CoroutineScope(Dispatchers.IO)
                 .launch {
-                    var newExpiration = com.example.ingredient.database.ExpirationDateData(text, "slice", "2022-04-09","2022-04-09")
+                    var newExpiration = com.example.ingredient.data.ExpirationDateData(text, "slice", "2022-04-09","2022-04-09")
                     db?.expirationDateDao()?.insert(newExpiration)
                 }
             var exp = ""
