+++ app/src/main/java/com/example/ingredient/src/search/SearchAdapter.kt
@@ -1,7 +1,6 @@
-package com.example.ingredient
+package com.example.ingredient.src.search
 
 import android.content.Context
-import android.content.Intent
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -9,8 +8,7 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.ingredient.fragment.PurchaseConfirmationDialogFragment
-import com.google.firebase.firestore.FirebaseFirestore
+import com.example.ingredient.R
 
 class SearchAdapter (private val recipeList: MutableList<Array<String>>)
     : RecyclerView.Adapter<SearchAdapter.ViewHolder>()
@@ -75,5 +73,5 @@ class SearchAdapter (private val recipeList: MutableList<Array<String>>)
         fun setItemClickListener(onItemClickListener: OnItemClickListener) {
             this.itemCLickListener = onItemClickListener
         }
-        private lateinit var itemCLickListener:OnItemClickListener
+        private lateinit var itemCLickListener: OnItemClickListener
     }
