+++ app/src/main/res/layout/fragment_dialog.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.DialogFragment">
+    tools:context=".src.DialogFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
