+++ app/src/main/res/layout/fragment_expiration_date.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragment.ExpirationDate">
+    tools:context=".src.ExpirationDate">
 
     <TextView
         android:id="@+id/expiration_date_text"
