+++ app/src/main/res/layout/fragment_food_book.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragment.FoodBook">
+    tools:context=".src.FoodBook">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
