+++ app/src/main/res/layout/fragment_note.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.Note">
+    tools:context=".src.Note">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
@@ -23,6 +23,4 @@
         android:lines="25"
         android:paddingStart="20dp"
         android:paddingEnd="20dp" />
-
-
 </FrameLayout>
